+++ .gitignore
@@ -55,3 +55,6 @@ buck-out/
 
 # Bundle artifact
 *.jsbundle
+
+# Expo
+.expo
