+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.mappsexpo",
+    package = "com.mapps",
 )
 
 android_resource(
     name = "res",
-    package = "com.mappsexpo",
+    package = "com.mapps",
     res = "src/main/res",
 )
