+++ android/app/src/main/java/com/mapps/MainActivity.java
@@ -1,4 +1,4 @@
-package com.mappsexpo;
+package com.mapps;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "MappsExpo";
+        return "Mapps";
     }
 
     @Override
