+++ android/settings.gradle
@@ -8,6 +8,6 @@ includeUnimodulesProjects()
 include ':react-native-gesture-handler'
 project(':react-native-gesture-handler').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-gesture-handler/android')
 
-rootProject.name = 'MappsExpo'
+rootProject.name = 'Mapps'
 
 include ':app'
