+++ ios/Mapps/Images.xcassets/AppIcon.appiconset/Contents.json
